% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spat_funcs.R
\name{convert.HDS.to.mcmc}
\alias{convert.HDS.to.mcmc}
\title{function to convert HierarchicalDS MCMC list vector (used in estimation) into an mcmc object (cf. coda package)}
\usage{
convert.HDS.to.mcmc(MCMC, N.hab.pois.par, N.hab.bern.par, n.ind.cov,
  n.cov.cols, Hab.pois.names, Hab.bern.names, Det.names, Cov.names,
  MisID.names, N.par.misID = NULL, misID.mat = NULL,
  fix.tau.nu = FALSE, misID = TRUE, spat.ind = TRUE,
  point.ind = TRUE)
}
\arguments{
\item{MCMC}{list vector providing MCMC samples for each parameter type}

\item{N.hab.pois.par}{see help for mcmc_ds.R}

\item{N.hab.bern.par}{see help for mcmc_ds.R}

\item{n.ind.cov}{see help for mcmc_ds.R}

\item{n.cov.cols}{number of columns in Cov.prior.parms matrix}

\item{Hab.pois.names}{see help for mcmc_ds.R}

\item{Hab.bern.names}{see help for mcmc_ds.R}

\item{Det.names}{see help for mcmc_ds.R}

\item{Cov.names}{see help for mcmc_ds.R}

\item{MisID.names}{see help for mcmc_ds.R}

\item{N.par.misID}{see help for mcmc_ds.R}

\item{misID.mat}{see help for mcmc_ds.R}

\item{fix.tau.nu}{see help for mcmc_ds.R}

\item{misID}{see help for mcmc_ds.R}

\item{spat.ind}{see help for mcmc_ds.R}

\item{point.ind}{see help for mcmc_ds.R}
}
\description{
function to convert HierarchicalDS MCMC list vector (used in estimation) into an mcmc object (cf. coda package)
}
\author{
Paul B. Conn
}
\concept{MCMC}
\concept{coda}
