\name{dens.plot}
\alias{dens.plot}
\title{function to plot kernel density estimates for mcmc objects}
\usage{
  dens.plot(MCMC, xaxt = TRUE)
}
\arguments{
  \item{MCMC}{an mcmc object with columns referencing
  different parameter types (column names are used for
  plotting labels)}

  \item{xaxt}{if TRUE (default), prints x-axis values}
}
\description{
  function to plot kernel density estimates for mcmc
  objects
}
\author{
  Paul B. Conn
}
\keyword{density,}
\keyword{ggplot2}
\keyword{kernel}
\keyword{MCMC,}

