% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get_confusion_array}
\alias{get_confusion_array}
\title{Fill confusion array - one confusion matrix for each individual (DEPRECATED)}
\usage{
get_confusion_array(Confuse, Cov = NULL, Beta, n.indiv, misID.mat,
  misID.formulas, symm = TRUE)
}
\arguments{
\item{Confuse}{An 3-dimensional array, with dimensions (# of individuals, # of rows in misID.mat, # of cols of misID.mat)}

\item{Cov}{Data frame including all covariates for the misclassification model (individuals are on rows)}

\item{Beta}{A list where each entry is a vector giving the parameters of the misID model}

\item{n.indiv}{Integer giving the number of individuals}

\item{misID.mat}{With true state on rows and assigned state on column, each positive entry provides an index to misID.models (i.e. what model to assume on multinomial logit space); a 0 indicates an impossible assigment; a negative number designates which column is to be obtained via subtraction}

\item{misID.formulas}{A formula vector providing linear model-type formulas for each positive value of misID.mat.  If the same model is used in multiple columns it is assumed that all fixed effects (except the intercept) are shared}

\item{symm}{if TRUE, symmetric classification probabilities are applied (e.g. pi^12=pi^21)}
}
\value{
A filled version of Confuse
}
\description{
Fill confusion array - one confusion matrix for each individual (DEPRECATED)
}
\author{
Paul B. Conn
}

