% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spat_funcs.R
\name{plot_obs_pred}
\alias{plot_obs_pred}
\title{plot 'observed' versus predicted values for abundance of each species at each transect}
\usage{
plot_obs_pred(Out)
}
\arguments{
\item{Out}{Output list from "mcmc_ds.R"}
}
\description{
plot 'observed' versus predicted values for abundance of each species at each transect
}
\author{
Paul B. Conn
}
