% Generated by roxygen2 (4.0.2): do not edit by hand
\name{post_loss}
\alias{post_loss}
\title{function to calculate posterior predictive loss given the output object from hierarchicalDS}
\usage{
post_loss(Out, burnin = 0)
}
\arguments{
\item{Out}{Output object from running hierarchicalDS}

\item{burnin}{Any additional #'s of values from beginning of chain to discard before calculating PPL statistic (default is 0)}
}
\value{
A matrix with posterior variance (P), sums of squares (G) for the posterior mean and median predictions (compared to Observations), and total posterior loss (D)
}
\description{
function to calculate posterior predictive loss given the output object from hierarchicalDS
}
\author{
Paul B. Conn
}
\keyword{Posterior}
\keyword{loss}
\keyword{predictive}

