\name{run_analysis}
\alias{run_analysis}
\title{function to simulate data and run a simple hierarchical example analysis using said dataset...}
\description{function to simulate data and run a simple hierarchical example analysis using said dataset}
\value{see help for hierarchical_DS}
\keyword{simulation}
\author{Paul B. Conn}
