\name{simdata}
\docType{data}
\alias{simdata}
\title{MCMC output from running example script in HierarchicalDS}
\description{
  MCMC output from running example in Hierarchical DS
}
\usage{simdata}
\format{A list, called "Out", containing MCMC output from HierarchicalDS}
\author{
  Paul Conn \email{paul.conn@noaa.gov}
}
\keyword{data}
\keyword{datasets}

