% Generated by roxygen2 (4.0.2): do not edit by hand
\name{stack_data_misID}
\alias{stack_data_misID}
\title{function to stack data for midID updates (going from four dimensional array to a two dimensional array including observed groups}
\usage{
stack_data_misID(Data, G.obs, g.tot.obs, n.Observers, n.transects, n.species,
  stacked.names, factor.ind)
}
\arguments{
\item{Data}{4-d dataset}

\item{G.obs}{matrix giving the total numer of groups observed at least once by species and transect}

\item{g.tot.obs}{total number of observations for animals seen at least once}

\item{n.Observers}{vector giving number of observers per transect}

\item{n.transects}{number of transects}

\item{n.species}{number of species}

\item{stacked.names}{column names for new stacked dataset}

\item{factor.ind}{a vector of indicator variables (1 = factor/categorical variable, 0 = continuous variable)}
}
\value{
a stacked dataset (in matrix form)
}
\description{
function to stack data for midID updates (going from four dimensional array to a two dimensional array including observed groups
}
\author{
Paul B. Conn
}
\keyword{data}
\keyword{stack}

