\name{summary_N}
\alias{summary_N}
\title{calculate parameter estimates and confidence intervals for various loss functions}
\usage{
  summary_N(Out)
}
\arguments{
  \item{Out}{Output list from "mcmc_ds.R"}
}
\value{
  summary.N list vector, with the first list index
  indicating species
}
\description{
  calculate parameter estimates and confidence intervals
  for various loss functions
}
\author{
  Paul B. Conn
}
\keyword{summary}

