\name{get_mod_matrix}
\alias{get_mod_matrix}
\title{function to produce a design matrix given...}
\usage{get_mod_matrix(Cur.dat, stacked.names, factor.ind, Det.formula, Levels)}
\description{function to produce a design matrix given}
\value{a design matrix}
\keyword{model}
\keyword{matrix}
\author{Paul B. Conn}
\arguments{\item{Cur.dat}{current dataset}
\item{stacked.names}{column names for current dataset}
\item{factor.ind}{a vector of indicator variables (1 = factor/categorical variable, 0 = continuous variable)}
\item{Det.formula}{a formula object}
\item{Levels}{A list object giving the number of levels for factor variables}}
