% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spat_funcs.R
\name{linear_adj}
\alias{linear_adj}
\title{Produce an adjacency matrix for a vector}
\usage{
linear_adj(x)
}
\arguments{
\item{x}{length of vector}
}
\value{
adjacency matrix
}
\description{
Produce an adjacency matrix for a vector
}
\author{
Paul Conn
}
