\name{linear_adj}
\alias{linear_adj}
\title{Produce an adjacency matrix for a vector}
\usage{
  linear_adj(x)
}
\arguments{
  \item{x}{length of vector}
}
\value{
  adjacency matrix
}
\description{
  Produce an adjacency matrix for a vector
}
\author{
  Paul Conn
}
\keyword{adjacency}

