\name{plot_obs_pred}
\alias{plot_obs_pred}
\title{plot 'observed' versus predicted values for abundance of each species at each transect}
\usage{
  plot_obs_pred(Out)
}
\arguments{
  \item{Out}{Output list from "mcmc_ds.R"}
}
\value{
  NULL
}
\description{
  plot 'observed' versus predicted values for abundance of
  each species at each transect
}
\author{
  Paul B. Conn
}
\keyword{diagnostics,}
\keyword{plot}

