% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plot_obs_pred}
\alias{plot_obs_pred}
\title{plot 'observed' versus predicted values for abundance of each species at each transect}
\usage{
plot_obs_pred(Out)
}
\arguments{
\item{Out}{Output list from "mcmc_ds.R"}
}
\description{
plot 'observed' versus predicted values for abundance of each species at each transect
}
\author{
Paul B. Conn
}
\keyword{diagnostics,}
\keyword{plot}

