% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spat_funcs.R
\docType{data}
\name{sim_out}
\alias{sim_out}
\title{MCMC output from running example in Hierarchical DS}
\description{
MCMC output from running example in Hierarchical DS
}
\author{
Paul Conn \email{paul.conn@noaa.gov}
}
