\name{simulate_data}
\alias{simulate_data}
\title{function to simulate distance sampling data from simple model with increasing abundance...}
\usage{simulate_data(S)}
\description{function to simulate distance sampling data from simple model with increasing abundance
intensity, no assumed spatial structure, and point independence}
\value{a distance sampling dataset}
\keyword{distance}
\keyword{sampling,}
\keyword{simulation}
\author{Paul B. Conn}
\arguments{\item{S}{number of spatial strata (a single transect is placed in each strata and assumed to cover the whole strata)}}
