% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spat_funcs.R
\name{square_adj}
\alias{square_adj}
\title{Produce an adjacency matrix for a square grid}
\usage{
square_adj(x)
}
\arguments{
\item{x}{number of cells on side of grid}
}
\value{
adjacency matrix
}
\description{
Produce an adjacency matrix for a square grid
}
\author{
Paul Conn
}
